+++ CHANGELOG.md
@@ -6,12 +6,14 @@
 - **command:** `command.execute` now returns `T | Promise<T>` (based on the execute function)
 - **command:** `command.execute` when error is thrown it will bubble the error
 - **command:** `Command` is now typed based on the execute function
+- **command:** `SsvCommand/SsvCommandRef` is now typed
 
 ### BREAKING CHANGES
 
 - **command:** remove `autoDestroy`, `destroy`, `subscribe`, `unsubscribe` from `ICommand`/`Command`
 - **command:** `commandAsync` deprecated use `command` instead
 - **command:** `Observable` `isExecuting` will change to false only when it completes (previously when it emits)
+- **command:** `SsvCommand/SsvCommandRef` strict typings
 
 ## 4.0.0 (2025-11-14)